### 2021-04-05 #3150 @whedon
[PRE REVIEW]: fitgrid: A Python package for multi-channel event-related time series regression modeling
**Submitting author:** @turbach (<a href="http://orcid.org/0000-0001-7993-142X">Thomas P. Urbach</a>)
**Repository:** <a href="https://github.com/kutaslab/fitgrid" target ="_blank">https://github.com/kutaslab/fitgrid</a>
**Version:** v0.5.0
