### 2021-04-05 #3150 @turbach
https://github.com/kutaslab/fitgrid/tree/joss-paper
