### 2021-04-14 0730497
add linux selector for meta.yaml compiler jinja
