+++ conda/meta.yaml
@@ -22,9 +22,9 @@ requirements:
     # 02/07/21 conda build OK w/ binutils_linux-64 8.2.0,  gcc/gxx_linux-64 7.3.0, _sysconfigdata_x86_64_conda_cos6_linux_gnu
     # 02/08/21 conda build fails w/ binutils_* 2.35, gxx_linux_64 9.3.0 _sysconfigdata_x86_64_conda_cos7_linux_gnu
     # - binutils_linux-64 !=2.35
-    # revision 03/12/21 drop binutils pin, restore compiler jinjas, py36, 37, 38 ok
-    - {{ compiler('c') }}
-    - {{ compiler('cxx') }}
+    # revision 03/12/21 drop binutils pin, restore compiler jinjas, py36, 37, 38 ok for linux
+    - {{ compiler('c') }}  # [linux]
+    - {{ compiler('cxx') }} # [linux]
   host:
     - python {{ python }}
     - pip
