### 2021-04-29 #3150 @meg-simula
@SRSteinkamp, @neurofractal, @ajquinn, @sappelhoff: You have been suggested as reviewers - would you be available and willing to review this package for JOSS? Your assistance and expertise would be much appreciated.
