### 2021-05-07 390ff9c
WIP switching CI to conda-forge priority w/ mkl
