### 2021-05-07 0860d31
CI start installs w/ mamba conda-forge strict priority
