### 2021-05-07 c91fec4
unskip CI numpy thread set test
