+++ .github/workflows/fitgrid-cid.yml
@@ -131,13 +131,13 @@ jobs:
       # create an env w/ local package
       - name: conda create env_${{ env.PY_VER }} python=${{ env.PY_VER }} ${{ env.PACKAGE_NAME}} and black
         run: |
-          black --check --verbose -S -l 79 .
           mamba create -n env_$PY_VER python=$PY_VER $PACKAGE_NAME "blas=*=mkl*" -c local -c ejolly # -c conda-forge -c defaults
           conda activate env_$PY_VER
-          mamba install -q black pytest pytest-cov
           conda list
           lscpu
           python -c 'import numpy; numpy.show_config()'
+          mamba install -q black pytest pytest-cov
+          black --check --verbose -S -l 79 .
 
       - name: pytest
         run: |
