+++ .github/workflows/fitgrid-cid.yml
@@ -137,7 +137,7 @@ jobs:
           lscpu
           python -c 'import numpy; numpy.show_config()'
           mamba install -q black pytest pytest-cov
-          black --check --verbose -S -l 79 .
+          black --diff --verbose -S -l 79 .
 
       - name: pytest
         run: |
