+++ .github/workflows/fitgrid-cid.yml
@@ -72,10 +72,11 @@ defaults:
 
 jobs:
   fitgrid-cid:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }} # ubuntu-latest
     strategy:
       matrix:
         py_ver: [3.7, 3.8]
+        os: [ubuntu-latest, macos-10.15]  # Intel macs
 
     env:
       PY_VER: ${{ matrix.py_ver }}
