### 2021-05-08 eafa044
patch ci miniconda install for osx
