### 2021-05-08 7b85fcb
update CI deploy for osx job matrix
