### 2021-05-08 2961519
drop -c defaults from conda mambabuild
