### 2021-05-08 a4e6a04
fix CI docstring notes
