### 2021-05-19 #3293 @whedon
[REVIEW]: fitgrid: A Python package for multi-channel event-related time series regression modeling
**Submitting author:** @turbach (<a href="http://orcid.org/0000-0001-7993-142X">Thomas P. Urbach</a>)
**Repository:** <a href="https://github.com/kutaslab/fitgrid" target ="_blank">https://github.com/kutaslab/fitgrid</a>
**Version:** v0.5.2

    - [x] questions on installation: https://github.com/kutaslab/fitgrid/issues/196
    - [x] https://github.com/kutaslab/fitgrid/issues/199
    - [x] https://github.com/kutaslab/fitgrid/issues/210

    - [x] https://github.com/kutaslab/fitgrid/issues/196
    - [x] https://github.com/kutaslab/fitgrid/issues/198
    - [x] need docs on how to run full test suite: https://github.com/kutaslab/fitgrid/issues/200
    - [x] https://github.com/kutaslab/fitgrid/issues/200

    - [x] some suggestions here: https://github.com/kutaslab/fitgrid/pull/195
    - [x] https://github.com/kutaslab/fitgrid/pull/195#issuecomment-849698067
