### 2021-05-19 #3293 @meg-simula
:wave:  @turbach @AJQuinn @sappelhoff  this is the review thread for the paper.

Both reviewers have checklists at the top of this thread with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines. We aim for reviews to be completed within about 2-4 weeks, but please let me know if any of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please feel free to ping me (@meg-simula) if you have any questions/concerns!
