### 2021-05-22 4653188
docs update for mamba, osx, dev version bump
