+++ fitgrid/__init__.py
@@ -10,7 +10,7 @@ from .io import (
 from .models import run_model, lm, lmer
 from . import utils, defaults
 
-__version__ = "0.5.1.dev3"
+__version__ = "0.5.1.dev4"
 
 # for use by pytests and docs
 DATA_DIR = Path(__file__).parent / "data"
