+++ docs/source/installation.rst
@@ -1,83 +1,265 @@
 .. _installation:
 
-============
+############
 Installation
-============
+############
 
-**TL;DR** Use `conda install` not `pip` and for real work, run
-`fitgrid` on a fast multicore x64_86 linux server with gobs of RAM.
+**TL;DR** Use `mamba` or `conda` to install the fitgrid conda package
+along with your other packages into a fresh conda environment on a
+fast multicore x64_86 computer with gobs of RAM.
 
 
------------------------------
-`conda install` (recommended)
------------------------------
+==================
+conda environments
+==================
 
-`fitgrid` is packaged and distributed on `Anaconda Cloud
+`fitgrid` is packaged on `anaconda.org
 <https://anaconda.org/kutaslab/fitgrid>`_ for installation into conda
-virtual environments using the `conda package manager
-<https://conda.io>`_. The virtual environment isolates the `fitgrid`
-installation to prevent clashes with what is already installed on your
-system and the `conda` installer automatically searches for compatible
-versions of the hundreds of Python and R packages `fitgrid` requires
-to run. The steps are simple but before proceeding it is prudent to
-have a general understanding of conda virtual environments and at
-least these commands: ``conda create ...``, ``conda activate ...``,
-``conda deactivate``, ``conda install -c ....``. See the `Conda Cheat
-Sheet
-<https://docs.conda.io/projects/conda/en/latest/user-guide/cheatsheet.html>`_
-for a summary.
+"virtual" environments using the (good) `conda <https://conda.io>`_ or
+(better) `mamba <https://mamba.readthedocs.io/en/latest/>`_ package
+manager. The virtual environment isolates the `fitgrid` installation
+to prevent clashes with what is already installed on your system.  The
+`conda` package installer automatically populates the environment with
+compatible versions of the hundreds of Python and R packages `fitgrid`
+requires to run. The `mamba` package installer does the same thing but
+much faster and, in some cases, more reliably.
 
-This example shows how to install `fitgrid` from a bash shell command
-prompt in a linux terminal window.  It assumes the `conda` executable
-is already installed (we use `miniconda3
-<https://docs.conda.io/en/latest/miniconda.html>`_). It bakes a date
-into the name and provisions the new environment with `fitgrid` and a
-compatible `jupyterlab`. The order of the `-c` conda channels matters,
-the channels are searched for compatible packages from left to right.
+The steps for creating conda environments and installing fitgrid are
+simple but it is prudent to first have a general understanding of
+conda virtual environments and at least these commands: ``conda create
+...``, ``conda activate ...``, ``conda deactivate``, ``conda
+install -c ....``. See the `Conda Cheat Sheet
+<https://docs.conda.io/projects/conda/en/latest/user-guide/cheatsheet.html>`_
+for a summary. For fine tuning and troubleshooting environments, it is
+useful to have a grasp of conda channels and channel priority and the
+difference between the default conda channels `main` and `r`
+commercially maintained by Anaconda and the parallel universe
+of the `conda-forge` channel maintained by the open-source community.
+
 
 .. _conda_install_fitgrid:
 
-**Example installation**
 
-.. code-block:: bash
+================================
+`mamba` install (best practices)
+================================
 
-    $ conda create --name fg_012021 -c kutaslab -c defaults -c conda-forge fitgrid jupyterlab
-    $ conda activate fg_012021
+These examples show how to install `fitgrid` from a bash shell command
+prompt in a linux or Mac terminal window.  They assume the `conda`
+executable is already installed (we use `miniconda3
+<https://docs.conda.io/en/latest/miniconda.html>`_ for development and
+testing) and `mamba` has been installed into the base conda
+environment with ``conda install --name base mamba``. The
[…]
