### 2021-05-26 eaf093c
add Python 3.9 to CI job matrix
