### 2021-05-28 kutaslab/fitgrid#196 @sappelhoff
[JOSS] questions about installation: conda channels, pip
(related to https://github.com/openjournals/joss-reviews/issues/3293)

Hi,

the installation requires many different `conda` channels. It seems to me like this could be simplified by porting all your packages (including fitgrid itself) to https://conda-forge.org/ ... it's possible that some conda packages that you depend on are already available from there.

Is that something you considered? What were the arguments for/against it?

Furthermore: Could you please elaborate why the `pip` installation is not officially supported? Is it because of interactions with R packages? Or are there other reasons? (In the paper you say that the pypi packages are not "routinely testest", ... is it because it's particularly hard/challenging to do, or because you didn't get around to doing it yet?)
