### 2021-06-03 7b661c6
update install docs
