+++ docs/source/installation.rst
@@ -17,12 +17,23 @@ conda environments
 <https://anaconda.org/kutaslab/fitgrid>`_ for installation into conda
 "virtual" environments using the (good) `conda <https://conda.io>`_ or
 (better) `mamba <https://mamba.readthedocs.io/en/latest/>`_ package
-manager. The virtual environment isolates the `fitgrid` installation
-to prevent clashes with what is already installed on your system.  The
-`conda` package installer automatically populates the environment with
-compatible versions of the hundreds of Python and R packages `fitgrid`
-requires to run. The `mamba` package installer does the same thing but
-much faster and, in some cases, more reliably.
+manager. A virtual environment isolates the `fitgrid` installation to
+prevent clashes with what is already installed elsewhere in your
+system and other virtual environments.  The `mamba` environment
+creator and package installer does the same thing as `conda` but much
+faster and, in some cases, more reliably. Either way, when the package
+manager installs fitgrid it also automatically populates the
+environment with compatible versions of the hundreds of Python and R
+packages `fitgrid` requires to run including `numpy
+<https://numpy.org/doc/stable>`_, `pandas
+<https://pandas.pydata.org/docs>`_, `matplotlib
+<https://matplotlib.org>`_, `pymer4 <https://eshinjolly.com/pymer4>`_,
+and `rpy2 <https://rpy2.github.io/doc/latest/html/index.html>`_ as
+well as `R <https://www.r-project.org/other-docs.html>`_, `lme4
+<https://cran.r-project.org/web/packages/lme4/index.html>`_, and
+`lmerTest
+<https://cran.r-project.org/web/packages/lmerTest/index.html>`_ to
+name a few. 
 
 The steps for creating conda environments and installing fitgrid are
 simple but it is prudent to first have a general understanding of
@@ -30,12 +41,52 @@ conda virtual environments and at least these commands: ``conda create
 ...``, ``conda activate ...``, ``conda deactivate``, ``conda
 install -c ....``. See the `Conda Cheat Sheet
 <https://docs.conda.io/projects/conda/en/latest/user-guide/cheatsheet.html>`_
-for a summary. For fine tuning and troubleshooting environments, it is
-useful to have a grasp of conda channels and channel priority and the
-difference between the default conda channels `main` and `r`
-commercially maintained by Anaconda and the parallel universe
-of the `conda-forge` channel maintained by the open-source community.
+for a summary.
 
+Our current recommended best practice for working with conda
+environments is to install the lightweight `miniconda3
+<https://docs.conda.io/en/latest/miniconda.html>`_ and then avoid
+polluting the "base" conda environment with data analysis and
+application packages like fitgrid.  Instead, create separate new
+working environments, each populated with the packages needed for a
+given project. The mamba package is an exception to this rule. If you
+elect to use mamba, follow the installation instructions carefully and
+install it into the **base** conda environment with ``conda
+install --name base mamba``.
+
+For fine tuning environments and working around incompatible package
+versions refer to the core `conda tasks
+<https://conda.io/projects/conda/en/latest/user-guide/tasks/index.html>`_
+especially `managing conda channels and channel priority
+<https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-channels.html>`_
+and `installing packages
+<https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-pkgs.html#installing-packages>`_.
+For working with mixed Python and R environments, it is especially
+important to attend to the difference between the defaults conda
+channels `main` and `r` where packages are commercially maintained by
+the Anaconda team and the somewhat parallel universe of the
+`conda-forge <https://conda-forge.org/>`_ channel where packages are
+maintained by the open-source community. When you intend to install
[…]
