### 2021-06-06 kutaslab/fitgrid#196 @sappelhoff
also, as discussed in a comment in (https://github.com/kutaslab/fitgrid/pull/195#discussion_r640678628) : installation instructions appear at different places and are currently inconsistent --> should be consistent.

Current spots (to check again later, once CI failures etc. have been fixed):

- [docs](https://kutaslab.github.io/fitgrid/installation.html)
- [README](https://github.com/kutaslab/fitgrid#recommended-for-full-installation)
- paper
