### 2021-06-06 kutaslab/fitgrid#198 @sappelhoff
[JOSS] improving docs by using intersphinx?
(related to https://github.com/openjournals/joss-reviews/issues/3293)

I noticed that several times throughout the docs you refer to methods, classes, etc. of other Python packages, such as `Pandas.DataFrame`.

Through the [intersphinx](https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html) sphinx extension, it is possible to have sphinx automatically render these references as links to those documentations. So that `Pandas.DataFrame` would lead to the pandas docs on DataFrame.

You can have a look of this in action here: https://mne.tools/mne-bids/stable/generated/mne_bids.write_raw_bids.html#mne_bids.write_raw_bids

This also extends to the API docs, so that parameter data types are also linked to what they are. That's particularly useful in cases like the following:

your own functions like `fitgrid.load_grid` sometimes have parameters like so:

`    grid – loaded FitGrid object`

With intersphinx, "FitGrid object" can become a link that leads users to the docs of FitGrid object, without having to search for it.

---

Overall using this feature will require some minimal syntactical changes in the docs and docstrings here and there, and some additions to docs/conf.py --> but I am pretty sure it'll be well worth it.
