### 2021-06-06 kutaslab/fitgrid#199 @sappelhoff
[JOSS] saving and loading grids: HDF5
(related to openjournals/joss-reviews#3293)

I stumbled over this warning on loading and saving "grids" in the [docs](https://kutaslab.github.io/fitgrid/user_guide.html#saving-and-loading-grids):

> Fitted grids are saved and loaded with Python pickle which is not guaranteed to be portable across different versions of Python. Unpickling unknown files is not secure (for details see the Python docs). Only load grids you trust such as those you saved yourself. For reproducibility and portability fit the grid, collect the results you need, and export the dataframe to a standard data interchange format.

Wouldn't it be relatively straight forward to write a function that takes a "grid", saves it to hdf5, and loads it back from hdf5 on demand? Given that hdf5 is already a dependency of this project, this seems like a low hanging fruit to me that would greatly improve usability of the project.
