### 2021-06-06 kutaslab/fitgrid#200 @sappelhoff
[JOSS] Contributing guidelines
(related to openjournals/joss-reviews#3293)

In the docs I found that you "welcome field reports" from Windows users. Other than that however, I did not see documentation on:

- in how far contributions to this project are welcome
- how to do them (a developer setup of the software, project specific issues/PR guidelines, etc.)
- which channel to use to ask for help or contact you: Just use GitHub issues?

I think that although some of these points may have an implicit answer (e.g., users should just ask on GitHub Issues), there is a benefit in spelling this out in a dedicated documentation.

I recommend to either create a `CONTRIBUTING.md` file at the root of the project, and linking to it from the docs and README, or to prepare a dedicated page in the docs, and link to it from the README (and the docs toc).
