### 2021-06-06 kutaslab/fitgrid#200 @sappelhoff
PS: The developer notes should also contain information on how to run the complete test suite.
