### 2021-06-06 #3293 @sappelhoff
I have made a first pass at the review and opened several issues and PRs that now need to be addressed (or reasonably rejected). My overall impression so far is, that fitgrid is a useful and well documented package that fills a niche. I'll make a final evaluation once all open points have been addressed by the authors (see above).
