### 2021-06-15 32d2dc0
test windows-latest CI w/ strict conda-forge
