+++ .github/workflows/fitgrid-cid.yml
@@ -77,8 +77,7 @@ jobs:
     strategy:
       matrix:
         py_ver: [3.7, 3.8]
-        # os: [ubuntu-latest, macos-10.15]  # Intel macs
-        os: [windows-latest]  # Intel macs
+        os: [ubuntu-latest, macos-10.15]  # Intel macs
 
     env:
       PY_VER: ${{ matrix.py_ver }}
