### 2021-06-16 2bb9a8c
juggle CI logging
