### 2021-06-18 kutaslab/fitgrid#203 @AJQuinn
[JOSS] Documentation and Tutorials
Hello - this is my first issue for the review in: https://github.com/openjournals/joss-reviews/issues/3293

I've been running through the documentation and exploring the package. Generally, I'm impressed with the software but have had some difficulties getting a fitgrid analysis running on some real data. I'd like to get this working before completing the review. I think the issue is mainly as the documentation is relatively brief at the moment.

General
- The API page seems to be incomplete, perhaps there is an error with autofunction? https://fitgrid.readthedocs.io/en/stable/reference.html
- I second the issue from the other reviewer about pip installation, conda is useful but a standalone pip installation can be much simpler if at all possible. I think it should be possible by adding `pymer4` to `install_requires` in `setup.py`, though the user would need to configure their R install manually.

Tutorial page
- The tutorial page is a reasonable help in getting started but could be improved by a real data example and some example figures and visualisations. The current 'continuous + categorical' examples are good generalisations but would be well supported by a worked example of some real data.
- The `fitgrid.epochs_from_dataframe` and `fitgrid.epochs_from_hdf` functions are presented as only taking a single positional argument but actually need four or five arguments respectively (https://github.com/kutaslab/fitgrid/blob/main/fitgrid/io.py#L14)

Data format and loading
- The data loading examples from dataframes and hdf5 functions seem very targeted at specific data formats (I'm guessing related to mkpy?). I tried to run an analysis on some of my own data but was unable to wrangle it into an epochs object. It would be useful to either include more examples on how an external user can load a dataset. For example how to load arbitrary data (eg a 3d numpy [epochs x sensors x time] array + covariates) or data from other common formats such as MNE-python Epochs objects. I've submitted a separate issue with an example of what I've tried https://github.com/kutaslab/fitgrid/issues/202 would be very useful to solve this so I can complete the review - perhaps I've missed something very simple! (If this sort of thing is out of scope for this project, that is fine but would be good to state explicitly somewhere)
- Related to this, could you provide a specification for valid dataframe and hdf5 inputs? or link to an existing one elsewhere?
