### 2021-06-19 kutaslab/fitgrid#196 @turbach
Hi @sappelhoff, and thanks, these are good questions. @andportnoy and I have considered them at length.

Yes, as you surmise pip installation is not supported because an important fitgrid feature depends on interoperating with R and R packages like lme4. This is also the answer to @AJQuinn r.e. why not pip in #203. It might make sense to remove all mention of pip since installing from pip is rarely what users need and developers can clone the repo and pip install from source.

And yes, we have watched as conda-forge has evolved in recent years and have considered distributing there from time to time.

Our current CI builds and installs the conda package from source, then runs pytests and generates the docs and sphinx-gallery examples with the just-built package as installed. If successful, the deploy phase uploads a stable release to conda channel label main and development release to label pre-release; the corresponding docs upload to their gh-pages. This is all done automatically, in one pass in a single github Action workflow [fitgrid-cid.yml](https://github.com/kutaslab/fitgrid/blob/fg051dev5/.github/workflows/fitgrid-cid.yml). The conda packaging is functionally equivalent to a conda-forge feedstock configured as in [CFEP 05: Providing pre-release builds on the conda-forge anaconda channel](https://github.com/conda-forge/cfep/blob/master/cfep-05.md) so yes, we could deploy on conda-forge. But this would mean dismantling the current tightly knit workflow and rebuilding the same functionality spread across two repos. No doubt doable with some thought and effort but to what end? There has to be a good reason to convert something simple into something more complicated and saving users a few keystrokes per occasional installation doesn't seem like much return on investment.

Furthermore, we have not picked sides in the split between anaconda default channels (stale-but-stable packages, affinity for Intel MKL builds of algebra libraries)  and conda-forge (fresh-but-churning packages, affinity for OpenBLAS). In our experience, the choice can have dramatic consequences (on our Xeon CPUs does the job finish Monday or Wednesday?) but much depends on the user's aims and hardware ([AMD Ryzen 3900X vs Intel Xeon 2175W Python numpy - MKL vs OpenBLAS](https://www.pugetsystems.com/labs/hpc/AMD-Ryzen-3900X-vs-Intel-Xeon-2175W-Python-numpy---MKL-vs-OpenBLAS-1560/)). Current guidance on the [conda-forge.org home page](https://conda-forge.org) is to configure .condarc defaults for strict conda-forge channel priority. Strict conda-forge priority is good practical advice for smoothly building environments populated predominantly with conda-forge packages but this may or may not be what every user wants or needs. Our view is that conda-forge priority should be an option for installing fitgrid, not a default.

Revisions:

The fitgrid CI workflow has been updated to test the package installed with conda-forge strict channel priority. As of this writing, the package also installs, as it has since inception, with (flexible) priority for the anaconda default channels together with conda-forge. This allows the user to choose. We are expanding the fitgrid online installation docs with additional examples illustrating both approaches, as well as examples showing how to tune the installation to enforce MKL or OpenBLAS builds of the algebra libraries. You can preview the new installation material online in the [fitgrid 0.5.1dev4 development version docs](https://kutaslab.github.io/fitgrid-dev-docs/installation.html#mamba-install-best-practices), additional docs updates for #198, #200, #203 are work in progress for 0.5.1.dev5 for inclusion in the stable v0.5.1 release.
