### 2021-06-19 kutaslab/fitgrid#203 @turbach
@AJQuinn thanks for the comments, my apologies, you may be referring to legacy readthedocs. We migrated to gh-pages. I forgot readthedocs was still up, will get with @andportnoy and remedy that.

If believe the fitgrid github repo homepage About panel and JOSS ms link to the [current stable v0.5.0 release docs](https://kutaslab.github.io/fitgrid) and you can also find links to stable and development version docs here: [kutsalab.github.io/fitgrid](https://kutaslab.github.io).

The current docs [QuickStart](https://kutaslab.github.io/fitgrid/auto_gallery/quickstart.html)  that shows how to get a plain vanilla pd.DataFrame into fitgrid and might help with wrangling the mne.Epochs.to_data_frame() for fitgrid. I use fitgrid with mne.Epochs and mne.Epochs.metadata myself, will follow up with specifics about that in my comments to #202.

The [development version docs](https://kutaslab.github.io/fitgrid-dev-docs/) are work in progress.

The current version (0.5.1.dev4) has has an expanded section on Installation r.e. #196 and I'll respond about pip and conda in comments there.

I've also made a start on updates in response to @sappelhoff #198, #200, #201 for the next version. Feel free to jump in if you see something that needs work.
