### 2021-06-21 kutaslab/fitgrid#196 @sappelhoff
Thanks for the explanation @turbach - your workflow makes sense to me, and I am looking forward to seeing the additional doc updates once they are finished (please ping me then).

> It might make sense to remove all mention of pip since installing from pip is rarely what users need and developers can clone the repo and pip install from source.

agreed
