### 2021-06-24 kutaslab/fitgrid#203 @AJQuinn
Hello - this makes a lot of sense! I think you're right, not sure how I got onto the old version. I'll close this issue and restart on the latest docs.

Thank you!
