### 2021-07-01 kutaslab/fitgrid#196 @AJQuinn
Hello - thanks for the detailed comments here, I also agree that the workflow above makes sense in this context.

A couple of additions would still be helpful for developers working from source. I was able to install fitgrid from source such that I could work with git. I managed to get it working but several additional dependencies needed to be added.

I would recommend expanding the package list in `install_requires` in setup.py and adding a requirements.txt file which contains a complete list of dependencies matched to the conda env. (there is some redundancy with the conda environment here - but if a developer should be pip installing from source these additions are helpful)

These don't necessarily need to be advertised to mainstream users in the docs but would be valuable to potential developers
