### 2021-07-02 #3293 @AJQuinn
Hello - I've completed my review of fitgrid, it works as described, has excellent testing coverage and clear documentation. I've highlighted a few points for improvement in issues on the main repo and would be happy to proceed once these are actioned.
