### 2021-07-06 9a27fda
updates for JOSS reviews
