+++ joss/paper.md
@@ -14,6 +14,7 @@ authors:
     orcid: 0000-0001-7993-142X
     affiliation: 1
   - name: Andrey S. Portnoy
+	orcid: 0000-0001-9997-9025
     affiliation: 1
 affiliations:
  - name: Department of Cognitive Science, University of California, San Diego
@@ -47,20 +48,21 @@ $\hat{\beta}_{0}(t)$ for the regression model $y(t) = \beta_{0}(t) +
 \varepsilon(t)$, fit by minimizing squared error [@SmiKut2015]. The
 average ERP can be viewed as a time series of model parameter
 estimates. Generalizing to more complex models such as multiple
-regression $y = \beta_{0} + \beta_{1}X_{1} + \ldots +
-\beta_{p}X_{p} + \varepsilon$, likewise produces time series of
-estimates for the constant and each regressor coefficient, the
-$\hat{\beta}_{0}(t), \hat{\beta}_{1}(t), \ldots, \hat{\beta}_{p}(t)$
-dubbed regression ERP (rERP) waveforms
+regression $y = \beta_{0} + \beta_{1}X_{1} + \ldots + \beta_{p}X_{p} +
+\varepsilon$, likewise produces time series of estimates for the
+constant and each regressor coefficient, the $\hat{\beta}_{0}(t),
+\hat{\beta}_{1}(t), \ldots, \hat{\beta}_{p}(t)$ dubbed regression ERP
+(rERP) waveforms
 [see @SmiKut2015; @SmiKut2015b for discussion of related approaches].
 This holds for straight-line fits ("slope" rERPs) as well as models of
 curvilinear relationships such as spline regression
 [@SmiKut2015b]. Besides the estimated coefficient rERPs, the approach
 also produces time series for all the basic and derived quantities of
-the fitted model: coefficient standard errors, residuals,
-likelihood, Akaike information criterion (AIC), and so
-forth. With the shift from averaging to regression modeling, however,
-comes a new problem.
+the fitted model: coefficient standard errors, residuals, likelihood,
+Akaike information criterion (AIC), and so forth. With the shift from
+averaging to regression modeling, however, comes a new problem:
+fitting, diagnosing, comparing, evaluating and interpreting large
+numbers of regression models.
 
 # Statement of need
 
@@ -82,27 +84,43 @@ samples per second, in 3 seconds of 32-channel EEG data there are
 candidate models requires 72,000 separate model fits, where the size
 of each data set might range anywhere from a few dozens of
 observations for a single subject to tens of thousands of observations
-for a large scale experiment. The combinatorial explosion is
-unavoidable; `fitgrid` contains it.
+for a large scale experiment.
+
+The combinatorial explosion of model fits is unavoidable; `fitgrid`
+contains it by gathering the rich fit objects in a regular Time x
+Channel grid that provides users with approachable access to
+corresponding grids of the fit objects' attributes. There are various
+Python implementations that use matrix operations operations to
+efficiently estimate ordinary linear regression coefficients *en
+masse* for N-D arrays of 1-D vectors such as
+[numpy.linalg.lstsq](https://numpy.org/doc/stable/reference/generated/numpy.linalg.lstsq.html)
+[@Harris2020] or
+[sklearn.linear_model.LinearRegression](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html)
+[@scikit-learn]. By capturing the `statsmodels` and `pymer4.Lmer`
+model fit objects on tidy grids, `fitgrid` provides users with equally
+approachable access to the estimated coefficients and, crucially, the
+other quantities baked into the fit result objects such as coefficient
+standard errors, residuals, likelihood, and goodness of fit measures
+required for evaluating, comparing, and interpreting models and,
[…]
