+++ joss/paper.bib
@@ -13,24 +13,32 @@
   type =	 {Journal Article}
 }
 
-@inbook{Berger1930,
-   author =     {Berger, Hans},
-   title =      {{Electroencephalogram of man}},
-   booktitle =  {Hans Berger on the Electroencephalogram of Man. The
-                Fourteen Original Reports on the Human
-                Electroencephalogram},
-   editor =     {Gloor, Pierre},
-   publisher =  {Elsevier},
-   address =    {1969. Electroencephalography and Clinical
-                Neurophysiology, supplement 28. New York},
-   year =       {1930},
-   type =       {Book Section}
+@software{conda,
+  title = {conda},
+  author = {Anaconda Incorporated},
+  year = {2012-2021},
+  type = {Computer Software}
 }
 
-@article{Dawson1951,
+@inbook{Berger1930,
+  author =	 {Berger, Hans},
+  title =	 {{Electroencephalogram of man}},
+  booktitle =	 {Hans Berger on the Electroencephalogram of Man. The
+                  Fourteen Original Reports on the Human
+                  Electroencephalogram},
+  series =	 {Electroencephalography and Clinical Neurophysiology,
+                  supplement 28.},
+  editor =	 {Gloor, Pierre},
+  publisher =	 {Elsevier},
+  address =	 {New York},
+  year =	 1930,
+  type =	 {Book Section}
+}
+
+@inproceedings{Dawson1951,
   author =	 {Dawson, G.~D.},
-  title =	 {{A summation technique for detecting small signals in
-                  a large irregular background}},
+  title =	 {{A summation technique for detecting small signals
+                  in a large irregular background}},
   journal =	 {Journal of Physiology},
   volume =	 115,
   number =	 1,
@@ -58,11 +66,13 @@
   author =	 {Delorme, A. and Makeig, S.},
   year =	 2004,
   title =	 {{EEGLAB: An open-source toolbox for analysis of
-                  single-trial EEG dynamics}},
+                  single-trial EEG dynamics including independent
+                  component analysis}},
   journal =	 {Journal of Neuroscience Methods},
   volume =	 134,
   pages =	 {9-21},
   type =	 {Journal Article},
+  doi =		 {10.1016/j.jneumeth.2003.10.009}
 }
 
 @article{EhiDim2019,
@@ -75,7 +85,8 @@
   pages =	 {e7838},
   year =	 2019,
   type =	 {Journal Article},
-  publisher =	 {PeerJ Inc.}
+  publisher =	 {PeerJ Inc.},
+  doi =		 {10.7717/peerj.7838}
 }
 
 @article{GramfortEtAl2013,
@@ -93,6 +104,46 @@
                   {https://www.frontiersin.org/article/10.3389/fnins.2013.00267},
   doi =		 {10.3389/fnins.2013.00267},
   issn =	 {1662-453X},
+}
+
+@article{         Harris2020,
+  title =	 {Array programming with {NumPy}},
+  author =	 {Charles R. Harris and K. Jarrod Millman and
+                  St{\'{e}}fan J.  van der Walt and Ralf Gommers and
+                  Pauli Virtanen and David Cournapeau and Eric Wieser
+                  and Julian Taylor and Sebastian Berg and Nathaniel
+                  J. Smith and Robert Kern and Matti Picus and Stephan
+                  Hoyer and Marten H. van Kerkwijk and Matthew Brett
+                  and Allan Haldane and Jaime Fern{\'{a}}ndez del
[…]
