### 2021-07-06 aa5b277
update docs for JOSS reviews
