+++ docs/source/gallery/workflow.py
@@ -3,13 +3,13 @@
 Workflow Outline
 ################
 
-**TL; DR** conda install `fitgrid` and `jupyter` into a fresh conda
+**TL; DR** conda install ``fitgrid`` and ``jupyter`` into a fresh conda
 virtual environment (shown :ref:`here
 <conda_install_fitgrid>`). Download this document as Jupyter notebook
 and run it. Read the notes, explore the output. Prerequisites: a
 working knowledge of conda and virtual environments, regression modeling
 with formulas like :math:`\mathsf{\sim ~ 1 + a + b + a:b}` in Python or R,
-and a bit of general purpose Python and `pandas`.
+and a bit of general purpose Python and ``pandas``.
 
 
 With multichannel digital recordings in hand, the basic `fitgrid`
@@ -21,18 +21,19 @@ modeling workflow is four steps:
    variables (numeric, string, boolean). See :ref:`epochs_data_format` for
    details.
 
-#. Load the data into a `fitgrid.Epochs` object for modeling.
+#. Load the data into a ``fitgrid`` :py:class:`Epochs <fitgrid.epochs.Epochs>` object for modeling.
 
 #. Specify and fit ordinary least squares or linear mixed-effects
-   model formula to populate the `FitGrid[times, channels]` object
-   with the fit results.  For OLS models use the `patsy` formula
-   syntax which works like `lm` formulas in R. For mixed-effects
-   models use the `lme4::lmer` R formula syntax.
+   model formula to populate the :py:class:`FitGrid[times, channels]
+   <fitgrid.fitgrid.FitGrid>` object with the fit results.  For OLS
+   models use the `patsy` formula syntax which works like `lm`
+   formulas in R. For mixed-effects models use the `lme4::lmer` R
+   formula syntax.
 
-#. Query and slice the `FitGrid[times, channels]` to fetch tidy time x
-   channel dataframes of model fit results: coefficient estimates,
-   residual error, model diagnostics, etc. for visualization and
-   analysis.
+#. Query and slice the :py:class:`FitGrid[times, channels]
+   <fitgrid.fitgrid.FitGrid>` to fetch tidy time x channel dataframes
+   of model fit results: coefficient estimates, residual error, model
+   diagnostics, etc. for visualization and analysis.
 
 """
 
@@ -40,9 +41,9 @@ modeling workflow is four steps:
 # 1. Prepare epochs data
 # ======================
 #
-# `fitgrid` assumes you are modeling epochs, i.e., fixed-length
+# ``fitgrid`` assumes you are modeling epochs, i.e., fixed-length
 # segments of typically multi-channel digital data, time stamped, and
-# collected in a tidy `pandas.DataFrame` (see
+# collected in a tidy :py:class:`pandas.DataFrame` (see
 # :ref:`epochs_data_format` for details).
 #
 # Here is a small complete simulated data set with four epochs and
@@ -70,7 +71,7 @@ epochs_df  # display
 
 # %%
 # A real epochs data set is typically much larger but the format is the same
-# as illustrated here with sample EEG data that `fitgrid` downloads from
+# as illustrated here with sample EEG data that ``fitgrid`` downloads from
 # `Zenodo <https://zenodo.org/record/4099632#.YAMXh5NKj6Mzenodo>`_.
 #
 # .. image:: ../_static/eeg_events.png
@@ -104,11 +105,11 @@ p3_epochs_df = pd.read_feather(DATA_DIR / "sub000p3.ms1500.epochs.feather")
 #
 # **Data wrangling.** As with any data analysis pipeline, the epochs
 # data must be quality controlled and groomed to final form before
-# modeling. `fitgrid` ingests tidy `pandas.DataFrames` directly which
[…]

[…]
