### 2021-07-06 kutaslab/fitgrid#199 @turbach
@sappelhof I'm tagging this as an enhancement, but I think there is more to it. I'll leave the issue open for discussion if there's interest.
