### 2021-07-06 3ee23e5
fix docs typos, add module docstrings, clobber setup.py install_packages
