+++ fitgrid/errors.py
@@ -1,2 +1,6 @@
+# -*- coding: utf-8 -*-
+"""Developer stub for internal fitgrid errors"""
+
+
 class FitGridError(Exception):
     pass
