+++ setup.py
@@ -1,12 +1,6 @@
-# to install in development mode, from this directory run
-#
-#     pip install --user -e .
-#
-# to install stable package systemwide, as root run
-#
-#     pip install .
-#
-# http://python-packaging.readthedocs.io/en/latest/minimal.html
+"""To install fitgrid in development mode see
+kutaslab.github.io/fitgrid-dev-docs/contributing.html
+"""
 
 import re
 from setuptools import find_packages, setup
@@ -14,6 +8,8 @@ from setuptools import find_packages, setup
 
 # from fitgrid.version import __version__
 def get_ver():
+    """format check"""
+
     with open("./fitgrid/__init__.py", "r") as stream:
         fg_ver = re.search(
             r".*__version__.*=.*[\"\'](?P<ver_str>\d+\.\d+\.\d+\S*)[\'\"].*",
@@ -21,21 +17,21 @@ def get_ver():
         )
 
     if fg_ver is None:
-        msg = f"""
+        msg = """
         fitgrid.__init__.py must have an X.Y.Z semantic version, e.g.,
-
         __version__ = '0.0.0'
         __version__ = '0.0.0-dev.0.0'
-
         """
         raise ValueError(msg)
-    else:
-        return fg_ver['ver_str']
+
+    return fg_ver['ver_str']
 
 
 def readme():
-    with open('README.md') as f:
-        return f.read()
+    """slurp text"""
+
+    with open('README.md') as strm:
+        return strm.read()
 
 
 setup(
@@ -53,12 +49,4 @@ setup(
         "Intended Audience :: Science/Research",
     ],
     packages=find_packages(exclude=['tests']),
-    install_requires=[
-        'patsy',
-        'statsmodels',
-        'matplotlib',
-        'scipy',
-        'tqdm',
-        'tables',
-    ],
 )
