+++ docs/source/contributing.rst
@@ -1,0 +1,313 @@
+.. _how_to_contribute:
+
+############
+Contributing
+############
+
+
+Field reports and ideas large and small for how to improve ``fitgrid``
+are welcome. Please post what you have in mind on GitHub in
+``fitgrid`` `Issues <https://github.com/kutaslab/fitgrid/issues>`_ in
+accordance with the ``fitgrid`` `Code of Conduct
+<https://github.com/kutaslab/fitgrid/blob/main/CODE_OF_CONDUCT.md>`_
+to start a discussion of next steps.
+
+The ``fitgrid`` code is written in Python, and requires many
+open-source scientific computing packages including :std:doc:`numpy
+<numpy:index>`, :std:doc:`pandas <pandas:index>`, :std:doc:`matplotlib
+<matplotlib:index>`, and :std:doc:`statsmodels
+<statsmodels:index>`. The linear mixed-effects modeling further
+requires the :std:doc:`pymer4 <pymer4:index>` and :std:doc:`rpy2
+<rpy2:index>` Python packages as well as the `R
+<https://www.r-project.org/other-docs.html>`_ language, and many R
+packages including `lme4
+<https://cran.r-project.org/web/packages/lme4/index.html>`_, and
+`lmerTest
+<https://cran.r-project.org/web/packages/lmerTest/index.html>`_. Version
+control is managed with git and the primary source repository is
+hosted on GitHub in https://github.com/kutaslab.
+
+The stable and pre-release :std:doc:`fitgrid <fitgrid:index>` packages
+are available on https://anaconda.org/kutaslab/fitgrid for easy
+installation into conda virtual environments with :std:doc:`conda
+<conda:index>` or :std:doc:`mamba <mamba:index>`. The Python package
+installer :std:doc:`pip <pip:index>` is not supported because of the R
+dependencies.
+
+The documentation is generated with :std:doc:`sphinx
+<sphinx:contents>` and :std:doc:`sphinx-gallery
+<sphinx-gallery:index>`. 
+
+=======================
+Overview for developers
+=======================
+
+Versions
+  ``fitgrid`` is semantically versioned following a simplified 
+  Python `PEP 440 <https://www.python.org/dev/peps/pep-0440>`_ scheme.
+
+  * **Syntax.** Legal ``fitgrid`` version strings have three required
+    numeric segments, ``M.N.P`` for major, minor, patch versions and
+    an optional ``.devX`` suffix with numeric ``X``. As
+    usual, the numeric values increment monotonically except that subordinate
+    segments reset to 0 when a superordinate segment
+    increments. Version strings match the regular expression,
+    ``\d+\.\d+\.\d+(\.dev\d+){0,1}``.
+
+  * **Semantics.** ``M.N.P`` designates a stable version and ``M.N.P.devX``
+    designates a development version.
+    
+  .. note::
+     In this scheme, the stable package version ``M.N.P`` sorts higher
+     than the development version ``M.N.P.devX``, for example,
+     ``0.5.1.`` > ``0.5.1.dev1`` > ``0.5.1.dev0`` > ``0.5.0``
+
+GitHub reserved branch names and version strings: https:github.com/kutsalab/fitgrid
+  * The branch `main` = the latest stable release, version ``M.N.P`` with tag ``vM.N.P``
+  * The branch `dev` = the latest development version ``M.N.P.devX``.
+  * Names and version strings for working branches other than `main`
+    and `dev` are not strictly defined in this scheme. However, the
+    natural version sequence increments the dev segment to
+    ``M.N.P.devX+1`` following the pre-release package upload of
+    ``M.N.P.devX`` and increments the patch and resets the dev
[…]
