### 2021-07-07 kutaslab/fitgrid#200 @sappelhoff
Thanks, I think this has been well addressed.
