### 2021-07-07 kutaslab/fitgrid#198 @sappelhoff
Thanks for adding these. I see some links working now, some others do not work (yet).

What doesn't work:

- see: https://kutaslab.github.io/fitgrid-dev-docs/quick_reference.html#model-running
- for example the `RHS (str)` parameter --> the `str` should be a link to the Python docs, like for example `int` in these docs: https://pyprep.readthedocs.io/en/latest/generated/pyprep.NoisyChannels.html#pyprep.NoisyChannels
- there are other cases like that, but I don't know why it doesn't work right now.

What works:

- the "return type" fields in the API docs
- several other links in the gallery examples
