### 2021-07-07 kutaslab/fitgrid#196 @sappelhoff
The current overhaul addresses my concerns and I am fine with the new "warning" to never use pip unless you need to install in developer mode. the instructions are clear to me, so I will close this issue. @AJQuinn please let me know if you want me to reopen it.
