### 2021-07-07 88d3eec
fix typos contributing.rst
