+++ docs/source/contributing.rst
@@ -62,14 +62,14 @@ Versions
      than the development version ``M.N.P.devX``, for example,
      ``0.5.1.`` > ``0.5.1.dev1`` > ``0.5.1.dev0`` > ``0.5.0``
 
-GitHub reserved branch names and version strings: https:github.com/kutsalab/fitgrid
+GitHub reserved branch names and version strings: https://github.com/kutsalab/fitgrid
   * The branch `main` = the latest stable release, version ``M.N.P`` with tag ``vM.N.P``
   * The branch `dev` = the latest development version ``M.N.P.devX``.
   * Names and version strings for working branches other than `main`
     and `dev` are not strictly defined in this scheme. However, the
-    natural version sequence increments the dev segment to
+    natural version sequence increments the ``devX`` segment to
     ``M.N.P.devX+1`` following the pre-release package upload of
-    ``M.N.P.devX`` and increments the patch and resets the dev
+    ``M.N.P.devX`` and increments the patch and resets the ``devX``
     segment to ``M.N.P+1.dev0`` following the stable release of
     ``vM.N.P``.
 
@@ -101,34 +101,35 @@ Sphinx and sphinx-gallery documentation
   latest development version ``M.N.P.devX`` is deployed to
   `gh-pages-dev
   <https://github.com/kutaslab/fitgrid-dev/tree/gh-pages-dev>`_ and
-  available online at https:/kutaslab.github.io/fitgrid-dev-docs.
+  available online at https://kutaslab.github.io/fitgrid-dev-docs.
   
 
 ===========================================
 Continuous Integration and Deployment (CID)
 ===========================================
 
-The ``fitgrid`` CID is implemented in a single-pass GitHub
-Action workflow, `figrid-cid.yml
-<https://github.com/kutaslab/fitgrid/blob/main/.github/workflows/fitgrid-cid.yml>`_.
-The continuous integration workflow is triggered by push, pull request
-and manual release events on GitHub. The deploy phase selectively
-uploads the conda packages and documentation for development version
-pre-releases and stable releases. This scheme allows conda or mamba
-installation of development versions and automatically synchronizes
-the stable release version string and source code across the GitHub
-repository at `github.com/kutaslab/fitgrid
-<https://github.com/kutaslab/fitgrid>`_, the conda packages at
-`anaconda.org/kutaslab/fitgrid <https://anaconda.org>`_ , the online
-`sphinx documentation <https:kutaslab.github.io/fitgrid>`_, and the
-Zenodo source code archive at `DOI 10.5281/zenodo.3581496
-<https://doi.org/10.5281/zenodo.3581496>`_.
-
 .. _cid-figure:
 
 .. figure:: _static/fitgrid_cid_scheme.png
 
    Continuous Integration and Deployment Scheme
+
+
+The ``fitgrid`` CID is implemented in a single-pass GitHub Action
+workflow, `figrid-cid.yml
+<https://github.com/kutaslab/fitgrid/blob/main/.github/workflows/fitgrid-cid.yml>`_.
+The continuous integration workflow is triggered by push, pull request
+and manual release events on GitHub. The deploy phase selectively
+uploads the conda packages and documentation for development version
+pre-releases and stable releases. This scheme allows conda or mamba
+installation of both stable and development versions and automatically
+synchronizes the stable release version string and source code across
+the GitHub repository at `github.com/kutaslab/fitgrid
+<https://github.com/kutaslab/fitgrid>`_, the conda packages at
+`anaconda.org/kutaslab/fitgrid <https://anaconda.org>`_ , the online
+`sphinx documentation <https:kutaslab.github.io/fitgrid>`_, and the
[…]
