### 2021-07-07 d7b239c
fix typos, drop devX for 0.5.1 stable release
