### 2021-07-08 dc32d22
update binder/environment.yml for 0.5.1
