+++ binder/environment.yml
@@ -1,4 +1,4 @@
-name: mkgpu1_fg050_dev1_py37
+name: fg051_binder
 channels:
   - ejolly
   - kutaslab/label/pre-release
@@ -6,201 +6,259 @@ channels:
   - defaults
 dependencies:
   - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=1_llvm
-  - _r-mutex=1.0.0=anacondar_1
-  - abseil-cpp=20200225.2=he1b5a44_2
-  - arrow-cpp=3.0.0=py37hced866c_0
-  - attrs=20.3.0=pyhd3eb1b0_0
-  - binutils_impl_linux-64=2.31.1=h6176602_1
-  - binutils_linux-64=2.31.1=h6176602_9
-  - blas=1.0=mkl
-  - blosc=1.21.0=h8c45485_0
-  - boost-cpp=1.73.0=h27cfd23_11
-  - brotli=1.0.9=he6710b0_2
-  - brotlipy=0.7.0=py37h27cfd23_1003
-  - bwidget=1.9.11=1
-  - bzip2=1.0.8=h7b6447c_0
-  - c-ares=1.17.1=h27cfd23_0
-  - ca-certificates=2021.1.19=h06a4308_1
-  - cairo=1.16.0=hf32fb01_1
-  - certifi=2020.12.5=py37h06a4308_0
-  - cffi=1.14.5=py37h261ae71_0
-  - chardet=4.0.0=py37h06a4308_1003
-  - cryptography=3.4.6=py37hd23ed53_0
-  - curl=7.75.0=h979ede3_0
-  - cycler=0.10.0=py37_0
-  - dbus=1.13.18=hb2f20db_0
+  - _openmp_mutex=4.5=1_gnu
+  - _r-mutex=1.0.1=anacondar_1
+  - abseil-cpp=20210324.2=h9c3ff4c_0
+  - alsa-lib=1.2.3=h516909a_0
+  - arrow-cpp=4.0.1=py38h4710c81_4_cpu
+  - aws-c-cal=0.5.11=h95a6274_0
+  - aws-c-common=0.6.2=h7f98852_0
+  - aws-c-event-stream=0.2.7=h3541f99_13
+  - aws-c-io=0.10.5=hfb6a706_0
+  - aws-checksums=0.1.11=ha31a3da_7
+  - aws-sdk-cpp=1.8.186=hb4091e7_3
+  - binutils_impl_linux-64=2.36.1=h193b22a_1
+  - binutils_linux-64=2.36=hf3e587d_31
+  - blosc=1.21.0=h9c3ff4c_0
+  - brotlipy=0.7.0=py38h497a2fe_1001
+  - bwidget=1.9.14=0
+  - bzip2=1.0.8=h7f98852_4
+  - c-ares=1.17.1=h7f98852_1
+  - ca-certificates=2021.5.30=ha878542_0
+  - cairo=1.16.0=h6cf1ce9_1008
+  - certifi=2021.5.30=py38h578d9bd_0
+  - cffi=1.14.5=py38ha65f79e_0
+  - chardet=4.0.0=py38h578d9bd_1
+  - cryptography=3.4.7=py38ha5dfef3_0
+  - curl=7.77.0=hea6ffbf_0
+  - cycler=0.10.0=py_2
+  - dbus=1.13.6=h48d8840_2
   - deepdish=0.3.6=pyh9f0ad1d_0
-  - double-conversion=3.1.5=he6710b0_1
[…]
