+++ binder/README.txt
@@ -4,13 +4,25 @@ rest of the repo, in particular environment.yml
 The environment.yml must have what is necessary to run fitgrid but not
 be so specific that it doesn't run on the binder platform.
 
+To get binder working ahead of a stable M.N.P release, make sure that
+the dev branch with M.N.P pushed fitgrid M.N.P (no devX) to
+kutaslab/fitgrid/pre-release.
+
+Then ...
+
 1. To capture a usable environment.yml this has worked
 
-Install a stable or pre-release version of fitgrid with into a
+Install a stable or pre-release version of fitgrid with conda and defaults priority into a
 Python 3.7 conda environment, activate the environment and dump it like so
 
-  conda create -n env_for_binder python=3.7 fitgrid -c kutaslab -c defaults -c conda-forge -c ejolly -y
-  activate env_for_binder
+  conda create -n fg_binder python=3.7 fitgrid -c kutaslab -c defaults -c conda-forge -c ejolly -y
+  activate fg_binder
+  conda env export -f environment.yml 
+
+Alternatively, in a Python 3.8 environment with mambma and conda-forge priority 
+
+  conda create -n fg_binder python=3.8 fitgrid -c kutaslab/label/pre-release -c ejolly -c conda-forge -y
+  activate fg_binder
   conda env export -f environment.yml
 
 If the current stable release dependencies are too far behind to run the binder notebook replace 
