+++ binder/environment.yml
@@ -10,7 +10,7 @@ dependencies:
   - _r-mutex=1.0.1=anacondar_1
   - abseil-cpp=20210324.2=h9c3ff4c_0
   - alsa-lib=1.2.3=h516909a_0
-  - arrow-cpp=4.0.1=py38h4710c81_4_cpu
+  - arrow-cpp=4.0.1=py37h3d702e5_4_cpu
   - aws-c-cal=0.5.11=h95a6274_0
   - aws-c-common=0.6.2=h7f98852_0
   - aws-c-event-stream=0.2.7=h3541f99_13
@@ -20,22 +20,22 @@ dependencies:
   - binutils_impl_linux-64=2.36.1=h193b22a_1
   - binutils_linux-64=2.36=hf3e587d_31
   - blosc=1.21.0=h9c3ff4c_0
-  - brotlipy=0.7.0=py38h497a2fe_1001
+  - brotlipy=0.7.0=py37hb5d75c8_1001
   - bwidget=1.9.14=0
   - bzip2=1.0.8=h7f98852_4
   - c-ares=1.17.1=h7f98852_1
   - ca-certificates=2021.5.30=ha878542_0
   - cairo=1.16.0=h6cf1ce9_1008
-  - certifi=2021.5.30=py38h578d9bd_0
-  - cffi=1.14.5=py38ha65f79e_0
-  - chardet=4.0.0=py38h578d9bd_1
-  - cryptography=3.4.7=py38ha5dfef3_0
+  - certifi=2021.5.30=py37h89c1867_0
+  - cffi=1.14.5=py37hc58025e_0
+  - chardet=4.0.0=py37h89c1867_1
+  - cryptography=3.4.7=py37h5d9358c_0
   - curl=7.77.0=hea6ffbf_0
   - cycler=0.10.0=py_2
   - dbus=1.13.6=h48d8840_2
   - deepdish=0.3.6=pyh9f0ad1d_0
   - expat=2.4.1=h9c3ff4c_0
-  - fitgrid=0.5.1=py38_gd7b239ca_0
+  - fitgrid=0.5.1=py37_gd7b239ca_0
   - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
   - font-ttf-inconsolata=3.000=h77eed37_0
   - font-ttf-source-code-pro=2.038=h77eed37_0
@@ -70,7 +70,7 @@ dependencies:
   - joblib=1.0.1=pyhd8ed1ab_0
   - jpeg=9d=h516909a_0
   - kernel-headers_linux-64=2.6.32=h77966d4_13
-  - kiwisolver=1.3.1=py38h1fd1430_1
+  - kiwisolver=1.3.1=py37h2527ec5_1
   - krb5=1.19.1=hcc1bbae_0
   - lcms2=2.12=hddcbb42_0
   - ld_impl_linux-64=2.36.1=hea4e1c9_1
@@ -118,130 +118,129 @@ dependencies:
   - lz4-c=1.9.3=h9c3ff4c_0
   - lzo=2.10=h516909a_1000
   - make=4.3=hd18ef5c_1
-  - markupsafe=2.0.1=py38h497a2fe_0
-  - matplotlib=3.4.2=py38h578d9bd_0
-  - matplotlib-base=3.4.2=py38hcc49a3a_0
-  - mock=4.0.3=py38h578d9bd_1
+  - markupsafe=2.0.1=py37h5e8e339_0
+  - matplotlib=3.4.2=py37h89c1867_0
[…]
