### 2021-07-08 #3293 @turbach
@sappelhoff @AJQuinn @andportnoy @meg-simula

Hi all, thanks much for the reviews, close reads, and constructive comments. The JOSS ms, fitgrid package, and docs are better for them.

The revised ms is available [above](https://github.com/openjournals/joss-reviews/issues/3293#issuecomment-876094423).

The revisions to the package and docs are incorporated into a new stable release, [fitgrid v0.5.1](https://github.com/kutaslab/fitgrid/releases/tag/v0.5.1).

The corresponding linux and OSX 0.5.1 conda packages are deployed on [anaconda.org/kutaslab/fitgrid](https://anaconda.org/kutaslab/fitgrid) and the 0.5.1 docs are on [kutaslab.github.io/fitgrid](https://kutaslab.github.io/fitgrid).

The revisions were substantial and I hope fixed more errors than they introduced. In addition to patching the broken CI, and numerous corrections, the major revisions to the docs are
* expanded citations
* intersphinx links to external docs
* a substantially updated [Installation](https://kutaslab.github.io/fitgrid/installation.html#how-to-install-fitgrid)  section with new examples of installation options for mamba vs. conda, conda-forge vs. default Anaconda channels, and MKL vs. OpenBLAS libraries.
* an entirely new section [Contributing](https://kutaslab.github.io/fitgrid/contributing.html) that includes
  * an [overview](https://kutaslab.github.io/fitgrid/contributing.html#overview-for-developers) of the version control and [continuous integration and deployment workflow](https://kutaslab.github.io/fitgrid/contributing.html#continuous-integration-and-deployment-cid) for developers.
  * [best practices](https://kutaslab.github.io/fitgrid/contributing.html#id1) for installing fitgrid from source in development mode into a compatible conda
  environment and running pytests and building the sphinx documentation.

Additional notes on specific issues and revisions are collected in this PR kutaslab/fitgrid#208

If further changes to the JOSS m.s. are in order please branch from and make the PR to [kutasab/fitgrid/joss-paper](https://github.com/kutaslab/fitgrid/tree/joss-paper).

If there are changes to the fitgrid code or docs please branch from and make the PR to [kutaslab/fitgrid/fg052dev0](https://github.com/kutaslab/fitgrid/tree/fg052dev0).

I'll be traveling and [off email](https://www.google.com/maps/place/Kearsarge+Pass/@36.7730151,-118.3767328,3a,75y,90t/data=!3m8!1e2!3m6!1sAF1QipODl4OvQ-pfWIcI5eYKOxFZysvGSdNXcy1jEwxI!2e10!3e12!6shttps:%2F%2Flh5.googleusercontent.com%2Fp%2FAF1QipODl4OvQ-pfWIcI5eYKOxFZysvGSdNXcy1jEwxI%3Dw114-h86-k-no!7i4032!8i3024!4m5!3m4!1s0x80bfc19c7376e487:0x56fb4963071f5781!8m2!3d36.7730151!4d-118.3767328) July 8 - 14, will respond to remaining and any new issues when I return.

Thanks again.

Tom
