### 2021-07-08 kutaslab/fitgrid#210 @sappelhoff
[JOSS] Running tests and building docs
(related to https://github.com/openjournals/joss-reviews/issues/3293)

Thanks for all your changes @turbach. I have stumbled over a (probably small) thing:

I installed fitgrid according to https://kutaslab.github.io/fitgrid/installation.html#fitgrid-stable-release, using `conda` with `-c conda-forge`

I then followed the CID script and installed test dependencies:

https://github.com/kutaslab/fitgrid/blob/5e917e6b149631006b63a654e28ff3a7900effbd/.github/workflows/fitgrid-cid.yml#L156

and docs dependencies:

https://github.com/kutaslab/fitgrid/blob/5e917e6b149631006b63a654e28ff3a7900effbd/.github/workflows/fitgrid-cid.yml#L169

When running `pytest` and `make -C docs html` I frequently got the following error:

`rpy2.robjects.packages.PackageNotInstalledError: The R package "lmerTest" is not installed.`

even after installing it via `conda install -c conda-forge r-lmertest`, I got a different error for the tests and for building the docs:

`AttributeError: module 'rpy2.robjects.pandas2ri' has no attribute 'ri2py'`

Is there (1) a real issue here, or (2) did I do something wrong?. In the case of (2): It'd be good to improve the docs for contributors on which additional packages to install to properly run the tests and build the docs.
