### 2021-07-08 kutaslab/fitgrid#210 @sappelhoff
one more thing: Running the test and docs build (or one of the two) downloaded some data, which are not git-ignored, and these show up in the git diff. A potential contributor may be confused by that and/or accidentally commit that data. If you can think of a way on how to avoid this (docs, better .gitignore, ...), that'd be great I think.
