### 2021-07-08 #3293 @sappelhoff
Thanks @turbach I have only three hopefully minor points that are left:

- [ ] https://github.com/kutaslab/fitgrid/issues/210 (this might be a problem with my machine and/or the docs, or an actual issue)
- [ ] https://github.com/kutaslab/fitgrid/issues/198 (feel free to ditch this one if you can't figure it out)
- [ ] https://github.com/kutaslab/fitgrid/pull/211 (minor PR to purge some unneeded sphinx build warnings)
