### 2021-07-15 kutaslab/fitgrid#210 @turbach
@sappelhoff right, we don't want sample data committed, the downloads will be gitignored in the next patch release.
