### 2021-07-15 kutaslab/fitgrid#210 @turbach
@sappelhoff thanks for the dev environment test drive. Your approach is generally sound and along the lines of the suggestions in the [How to contribute](https://kutaslab.github.io/fitgrid/contributing.html#id1) docs. I don't think it's a serious issue tho to diagnose I'd probably want to look at your .condarc, an MRE that includes the `conda activate ...` commands and a `conda list` dump to see what version of what got installed from which channel into the environment that was active for the pytest and sphinx docs build. I'm adding a bug report .github/ISSUE_TEMPLATE to the next `fitgrid` patch release to help spell this out.

But. All of this is moot at the moment because ...

While I was out last week, mamba [0.15.0](https://github.com/mamba-org/mamba/blob/master/CHANGELOG.md) dropped and changed the semantics of strict channel priority.

I expect this is a breaking change for the `mamba` installation procedures in the current `fitgrid` v0.5.1 docs and is definitely a [breaking change](https://github.com/kutaslab/fitgrid/actions/runs/1021125327) for our CI workflow and I guess others b.c. of mamba-org/boa#165. Apparently fixes are in the works, when the dust settles I'll revise the `fitgrid` CI and docs accordingly and @ you to wrap up this issue.
