### 2021-07-15 kutaslab/fitgrid#198 @turbach
@sappelhoff yes, I see where the `pyprep` and the MNE Python docs link parameters back to the Python docs. On the other hand, the `numpy`, `pandas`, or `matplotlib` docs don't do this: [numpy.ndarray](https://numpy.org/doc/stable/reference/generated/numpy.ndarray.html),  [pandas.DataFrame](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html), [matplotlib.figure.Figure](https://matplotlib.org/stable/api/figure_api.html?highlight=matplotlib%20figure%20figure#matplotlib.figure.Figure)

I believe the `fitgrid` docs can be built either way via the `napoleon_use_param` boolean in the sphinx [conf.py](https://github.com/kutaslab/fitgrid/blob/main/docs/source/conf.py) and it is currently set to `False` in line with `numpy`, `pandas`, and `matplotlib`.

So my take is everything works as intended, the issue is about the docs style. I don't have a strong opinion ... don't see what harm the links do (so why not?) or what use they are in practice (so why?).

@andportnoy any thoughts?
