+++ docs/source/api_docs/modules.rst
@@ -1,5 +1,3 @@
-:orphan:
-
 fitgrid
 =======
 
