+++ .github/workflows/fitgrid-cid.yml
@@ -120,7 +120,7 @@ jobs:
           # switch to mamba/boa and conda-forge
           conda config --add channels conda-forge
           # conda config --set channel_priority strict  # new mamba channel priority semantics in 0.15.0
-          conda install -n base -q conda-build conda-verify anaconda-client twine "mamba<0.15" "boa<0.6"
+          conda install -n base -q conda-build conda-verify anaconda-client twine mamba boa
           echo "# ------------------------------------------------------------"
           conda info -a
           echo "# ------------------------------------------------------------"
