### 2021-07-16 876f671
CI wrangling for mamba 0.15.1, boa 0.6.1
