### 2021-07-16 8500c74
patch CI, replace -c local w/ $CONDA_BLD_PATH
