### 2021-07-16 kutaslab/fitgrid#198 @sappelhoff
> So my take is everything works as intended, the issue is about the docs style. I don't have a strong opinion ... don't see what harm the links do (so why not?) or what use they are in practice (so why?).

I see, thanks for looking into this! I'll check this off on my JOSS list then, and you can feel free to close it once you either decide to (1) keep it how it is, or (2) implement it differently.
