### 2021-07-16 #3293 @sappelhoff
@meg-simula all my points have been addressed satisfactorily by @turbach and I am happy to recommend fitgrid for publication in JOSS.

Some issues I raised are still waiting to be resolved, but I don't consider them blockers for publication.
