### 2021-07-16 #3293 @AJQuinn
Hello - agreed, I'm happy that the changes in https://github.com/kutaslab/fitgrid/pull/208 address my points. Can recommend for publication once these are merged!
