### 2021-07-16 dcd8992
update install/contributing docs for mamba 0.15.1
