+++ docs/source/contributing.rst
@@ -8,14 +8,16 @@ Contributing
 Field reports and ideas large and small for how to improve ``fitgrid``
 are welcome. Please post what you have in mind on GitHub in
 ``fitgrid`` `Issues <https://github.com/kutaslab/fitgrid/issues>`_ in
-accordance with the ``fitgrid`` `Code of Conduct
+accord with the `Code of Conduct
 <https://github.com/kutaslab/fitgrid/blob/main/CODE_OF_CONDUCT.md>`_
-to start a discussion of next steps and plan the approach.
+to start a discussion of next steps and plan the approach. If you
+think you have encountered a bug, please follow the bug report
+guidlines when
 
 
-================
-Package overview
-================
+========
+Overview
+========
 
 The ``fitgrid`` code is written in Python, and requires many
 open-source scientific computing packages including :std:doc:`numpy
@@ -63,9 +65,11 @@ Versions
   .. note::
      In this scheme, the stable package version ``M.N.P`` sorts higher
      than the development version ``M.N.P.devX``, for example,
-     ``0.5.1.`` > ``0.5.1.dev1`` > ``0.5.1.dev0`` > ``0.5.0``
+     ``0.5.1`` > ``0.5.1.dev1`` > ``0.5.1.dev0`` > ``0.5.0``
 
-GitHub reserved branch names and version strings: https://github.com/kutsalab/fitgrid
+GitHub reserved branches and version strings: https://github.com/kutsalab/fitgrid
+  * The `main` and `dev` branches are reserved for deploying package
+    releases, only maintainers make pull requests to these branches.
   * The branch `main` = the latest stable release, version ``M.N.P`` with tag ``vM.N.P``
   * The branch `dev` = the latest development version ``M.N.P.devX``.
   * Names and version strings for working branches other than `main`
@@ -107,44 +111,44 @@ Sphinx and sphinx-gallery documentation
   available online at https://kutaslab.github.io/fitgrid-dev-docs.
   
 
-The ``fitgrid`` CID is implemented in a single-pass GitHub Action
-workflow, `figrid-cid.yml
-<https://github.com/kutaslab/fitgrid/blob/main/.github/workflows/fitgrid-cid.yml>`_.
-The continuous integration workflow is triggered by push, pull request
-and manual release events on GitHub. The deploy phase selectively
-uploads the conda packages and documentation for development version
-pre-releases and stable releases. This scheme allows conda or mamba
-installation of both stable and development versions and automatically
-synchronizes the stable release version string and source code across
-the GitHub repository at `github.com/kutaslab/fitgrid
-<https://github.com/kutaslab/fitgrid>`_, the conda packages at
-`anaconda.org/kutaslab/fitgrid <https://anaconda.org>`_ , the online
-`sphinx documentation <https:kutaslab.github.io/fitgrid>`_, and the
-Zenodo source code archive at `DOI 10.5281/zenodo.3581496
-<https://doi.org/10.5281/zenodo.3581496>`_.
+Continuous Integration and Deployment (CID)
+  The ``fitgrid`` CID is implemented in a single-pass GitHub Action
+  workflow, `figrid-cid.yml
+  <https://github.com/kutaslab/fitgrid/blob/main/.github/workflows/fitgrid-cid.yml>`_.
+  The continuous integration workflow is triggered by push, pull
+  request and manual release events on GitHub. The deploy phase
+  selectively uploads the conda packages and documentation for
+  development version pre-releases and stable releases. This scheme
+  allows conda or mamba installation of both stable and development
+  versions and automatically synchronizes the stable release version
+  string and source code across the GitHub repository at
+  `github.com/kutaslab/fitgrid
[…]
