+++ docs/Makefile
@@ -21,7 +21,8 @@ help:
 
 api-rst: 
 	sphinx-apidoc -e -f -o source/api_docs ../fitgrid
-	rm source/api_docs/modules.rst
+	awk 'BEGIN{printf(":orphan:\n\n")}{print}' source/api_docs/modules.rst > source/api_docs/modules.tmp
+	mv source/api_docs/modules.tmp source/api_docs/modules.rst
 	ls source/api_docs
 
 # just format .rst, don't run sphinx-gallery scripts
