### 2021-07-17 eef3185
version bump for release 0.5.2, drop .dev0
