### 2021-07-18 c2ef81b
update install, contributing docs for mamba 0.15
