### 2021-07-18 22f8f39
version bump for 0.5.2, drop .dev1
