### 2021-07-19 #3293 @turbach
Hi @arfon, thanks much for stepping in.

Our software source is on [github.com/kutaslab/fitgrid](https://github.com/kutaslab/fitgrid) and we tag and release [stable versions](https://github.com/kutaslab/fitgrid/releases), most recently v0.5.1 and v0.5.2.

The changes from the JOSS reviews were already included in fitgrid v0.5.1 released on July 8. As luck would have it, on July 9 mamba 0.15.0 was released with a change in [strict channel priority](https://github.com/mamba-org/mamba/blob/master/CHANGELOG.md#0150-july-9-2021) semantics so there's a new fitgrid v0.5.2 release to patch up our [continuous integration](https://github.com/kutaslab/fitgrid/actions/runs/1043622628) and [online docs](https://kutaslab.github.io/fitgrid) for that.

As for the software archive, since late 2019 we have had [Zenodo DOI 10.5281/zenodo.3581496](https://doi.org/10.5281/zenodo.3581496) linked to the fitgrid github repo so it automatically archives each github release. This is the same Zenodo DOI we note in the JOSS submission on p. 4 line 136. It is a complete archive of the fitgrid stable releases from v0.4.7 through v0.5.1 and v0.5.2 and it names both of us JOSS submission authors, @andportnoy and myself (and also now @sappelhoff b.c. his recent PRs). Since we already have an established source archive in place, could you please take a look and see if it is acceptable for present purposes? Thanks.
